# apiDoc.yaml
swagger: "2.0"
info:
  version: "1.0.0"
  title: myDeben
  description: "API Documentation"
  contact:
    email: "youssuf.radi@deemalab.com"
host: "localhost:3000"
basePath: "/api"
...
schemes:
- "http"
paths:
...
  /store/generate/code:
    get:
      tags:
      - "Store"
      summary: "Returns QR code as a PNG for hotel admin representing his store"
      description: "A hotel manager generates QR code for users to scan"
      operationId: "storeQR"
      produces:
      - "application/json"
      security:
        - ApiKeyAuth: [store]
      responses:
        200:
          description: "QR image string base64"
          schema:
            type: "object"
            properties:
              code:
                type: "string"
                format: "base64"
        400:
          description: "Wrong Store token Used"
        401:
          description: "Token expired or Malformed"
        403:
          description: "Permission Denied"
  /store/generate/token/{ref}:
    get:
      tags:
      - "Store"
      summary: "Returns Token for hotel admin representing his store and a reference number"
      description: "A hotel manager generates token for users to enter and check in with"
      operationId: "storeToken"
      produces:
      - "application/json"
      parameters:
      - in: "path"
        name: "ref"
        description: "Refrence number to track user with latter"
        required: true
        type: "string"
      security:
        - ApiKeyAuth: [store]
      responses:
        200:
          description: "QR image string base64"
          schema:
            type: "object"
            properties:
              code:
                type: "string"
                format: "base64"
        400:
          description: "Wrong Store token Used"
        401:
          description: "Token expired or Malformed"
        403:
          description: "Permission Denied"
  /store/category:
    get:
      tags:
      - "Store"
      summary: "Retrives all store categories"
      description: "Shows all available categories for the store with all details"
      operationId: "storeShowCategories"
      produces:
      - "application/json"
      security:
        - ApiKeyAuth: [store]
      responses:
        200:
          description: "Array of categories"
          schema:
            type: "object"
            properties:
              categories:
                type: "array"
                items:
                  $ref: "#/definitions/Category"
        400:
          description: "Wrong Store token Used"
        401:
          description: "Token expired or Malformed"
        403:
          description: "Permission Denied"
    post:
      tags:
      - "Store"
      summary: "Adds a category to a store"
      description: "Adds a new category to the store with all it's details"
      operationId: "storeAddCategories"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "Required Fields for a category"
        required: true
        schema:
          type: "object"
          required:
          - "name"
          properties:
            name:
              type: "string" 
            picture:
              type: "string"
            description:
              type: "string"
      security:
        - ApiKeyAuth: [store]
      responses:
        200:
          description: "Catgory Added Successfully"
          schema:
            $ref: "#/definitions/Success"
        400:
          description: "Wrong User token Used or store doesnot exist"
        401:
          description: "Token expired or Malformed"
        403:
          description: "Permission Denied"
  /store/category/{categoryId}:
    patch:
      tags:
      - "Store"
      summary: "Edits category details in a store"
      description: "Edits category in the store with all it's details"
      operationId: "storeEditCategories"
      produces:
      - "application/json"
      parameters:
      - name: "categoryId"
        in: "path"
        description: "Category id to edit"
        required: true
        type: "integer"
        format: "int64"
      - in: "body"
        name: "body"
        description: "Optional Fields to edit a category"
        required: true
        schema:
          type: "object"
          properties:
            name:
              type: "string" 
            picture:
              type: "string"
            description:
              type: "string"
      security:
        - ApiKeyAuth: [store]
      responses:
        204:
          description: "Catgory Edited Successfully"
        400:
          description: "Wrong Store token Used "
        401:
          description: "Token expired or Malformed"
        403:
          description: "Permission Denied"
        404:
          description: "Product doesnot exist"
    delete:
      tags:
      - "Store"
      summary: "Deletes a category in a store"
      operationId: "storeDeleteCategories"
      produces:
      - "application/json"
      parameters:
      - name: "categoryId"
        in: "path"
        description: "Category id to delete"
        required: true
        type: "integer"
        format: "int64"
      security:
        - ApiKeyAuth: [store]
      responses:
        204:
          description: "Catgory Deleted Successfully"
        400:
          description: "Wrong Store token Used "
        401:
          description: "Token expired or Malformed"
        403:
          description: "Permission Denied"
        404:
          description: "Product doesnot exist"
  /store/product:
    get:
      tags:
      - "Store"
      summary: "Retrives all store category products"
      description: "Shows all available category products for the store with all details"
      operationId: "storeShowProducts"
      produces:
      - "application/json"
      parameters:
      - name: "categoryId"
        in: "query"
        description: "Category Id to retreive products inside"
        required: true
        type: "string"
      security:
        - ApiKeyAuth: [store]
      responses:
        200:
          description: "Array of Products"
          schema:
            type: "object"
            properties:
              products:
                type: "array"
                items:
                  $ref: "#/definitions/Product"
        400:
          description: "Wrong Store token Used"
        401:
          description: "Token expired or Malformed"
        403:
          description: "Permission Denied"
        404:
          description: "Category Id not found"
    post:
      tags:
      - "Store"
      summary: "Adds a Product to a Category in the store"
      description: "Adds a new product to the category in the store with all it's details"
      operationId: "storeAddProducts"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "Required Fields for a Product"
        required: true
        schema:
          type: "object"
          required:
          - "name"
          - "price"
          - "category_id"          
          properties:
            name:
              type: "string" 
            picture:
              type: "string"
            description:
              type: "string"
            price:
              type: "integer"
            gem:
              type: "boolean"
      security:
        - ApiKeyAuth: [store]
      responses:
        200:
          description: "Product Added Successfully"
          schema:
            $ref: "#/definitions/Success"
        400:
          description: "Wrong Store token Used"
        401:
          description: "Token expired or Malformed"
        403:
          description: "Permission Denied"
        404:
          description: "Category Id not found"
  /store/product/{productId}:
    patch:
      tags:
      - "Store"
      summary: "Edits Product details in a store"
      description: "Edits Product in the store with all it's details"
      operationId: "storeEditProducts"
      produces:
      - "application/json"
      parameters:
      - name: "productId"
        in: "path"
        description: "Product id to edit"
        required: true
        type: "integer"
        format: "int64"
      - in: "body"
        name: "body"
        description: "Required Fields for a Product"
        required: true
        schema:
          type: "object"
          properties:
            name:
              type: "string" 
            picture:
              type: "string"
            description:
              type: "string"
            price:
              type: "integer"
            gem:
              type: "boolean"
      security:
        - ApiKeyAuth: [store]
      responses:
        204:
          description: "Product Edited Successfully"
        400:
          description: "Wrong Store token Used"
        401:
          description: "Token expired or Malformed"
        403:
          description: "Permission Denied"
        404:
          description: "Product doesnot exist"
    delete:
      tags:
      - "Store"
      summary: "Deletes a Product in a store"
      operationId: "storeDeleteProduct"
      produces:
      - "application/json"
      parameters:
      - name: "productId"
        in: "path"
        description: "Product id to delete"
        required: true
        type: "integer"
        format: "int64"
      security:
        - ApiKeyAuth: [store]
      responses:
        204:
          description: "Product Deleted Successfully"
        400:
          description: "Wrong Store token Used"
        401:
          description: "Token expired or Malformed"
        403:
          description: "Permission Denied"
        404:
          description: "Product doesnot exist"
  /store/order/current:
    get:
      tags:
      - "Store"
      summary: "Returns current orders Only made to the store"
      description: "Store could see current orders that need to be servered to users or cancel them"
      operationId: "storeCurrentOrders"
      produces:
      - "application/json"
      parameters:
      - name: "sort"
        in: "query"
        description: "Sort parameter to return sorted results could be 'user', 'category or 'date"
        required: false
        type: "string"
        enum:
        - "user"
        - "category"
        - "date"
      security:
        - ApiKeyAuth: [store]
      responses:
        200:
          description: "Current orders"
          schema:
            type: "object"
            properties:
              oredrs:
                type: "array"
                items:
                  $ref: "#/definitions/Order"
        400:
          description: "Wrong Store token Used"
        401:
          description: "Token expired or Malformed"
        403:
          description: "Permission Denied"
  /store/order:
    get:
      tags:
      - "Store"
      summary: "Returns all orders made to the store"
      description: "Store could see all the order history by all the users"
      operationId: "storeOrders"
      produces:
      - "application/json"
      parameters:
      - name: "sort"
        in: "query"
        description: "Sort parameter to return sorted results could be 'user', 'category or 'date"
        required: false
        type: "string"
        enum:
        - "user"
        - "category"
        - "date"
      security:
        - ApiKeyAuth: [store]
      responses:
        200:
          description: "All orders"
          schema:
            type: "object"
            properties:
              oredrs:
                type: "array"
                items:
                  $ref: "#/definitions/Order"
        400:
          description: "Wrong Store token Used"
        401:
          description: "Token expired or Malformed"
        403:
          description: "Permission Denied"
  /store/order/serve/{orderId}:
    get:
      tags:
      - "Store"
      summary: "Serves a pending order"
      description: "Store can serve a pending order to the user"
      operationId: "storeServeOrder"
      produces:
      - "application/json"
      parameters:
      - name: "orderId"
        in: "path"
        description: "Order Id to serve"
        required: true
        type: "string"
      security:
        - ApiKeyAuth: [store]
      responses:
        200:
          description: "Order Served"
          schema:
            $ref: "#/definitions/Success"
        400:
          description: "Wrong Store token Used"
        401:
          description: "Token expired or Malformed"
        403:
          description: "Permission Denied / Order already served / Order already cancelled / Old order"
  /store/order/cancel/{orderId}:
    get:
      tags:
      - "Store"
      summary: "Cancels a pending order"
      description: "Store could cancel pending orders"
      operationId: "storeCancelOrder"
      produces:
      - "application/json"
      parameters:
      - name: "orderId"
        in: "path"
        description: "Order Id to cancel"
        required: true
        type: "string"
      security:
        - ApiKeyAuth: [store]
      responses:
        200:
          description: "Order Cancelled"
          schema:
            $ref: "#/definitions/Success"
        400:
          description: "Wrong Store token Used"
        401:
          description: "Token expired or Malformed"
        403:
          description: "Permission Denied / Order already served / Order already cancelled / Old order"
  /store/users:
    get:
      tags:
      - "Store"
      summary: "Returns list of Checked in users in stores"
      operationId: "storeCheckedInUsers"
      produces:
      - "application/json"
      security:
        - ApiKeyAuth: [store]
      responses:
        200:
          description: "Users Checked In Store"
          schema:
          type: "object"
          properties:
            users:
              type: "array"
              items:
                type: "object"
                properties:
                  id:
                    type: "integer"
                  email:
                    type: "string" 
                    format: "email"
                  name:
                    type: "string"
        400:
          description: "Wrong Store token Used"
        401:
          description: "Token expired or Malformed"
        403:
          description: "Permission Denied"
  /store/checkout:
    get:
      tags:
      - "Store"
      summary: "Checks out user from a store"
      operationId: "storeCheckOut"
      produces:
      - "application/json"
      parameters:
      - name: "userId"
        in: "query"
        description: "User Id to check out"
        required: true
        type: "string"
      security:
        - ApiKeyAuth: [store]
      responses:
        200:
          description: "User Checked Out Successfully"
          schema:
            $ref: "#/definitions/Success"
        400:
          description: "Wrong Store token Used"
        401:
          description: "Token expired or Malformed"
        403:
          description: "Permission Denied"
definitions:
  Success:
    type: "object"
    properties:
      detail:
        type: "string" 
      success:
        type: "boolean"
...
  Order:
    type: "object"
    properties:
      id:
        type: "string"
      name:
        type: "string" 
      picture:
        type: "string"
      description:
        type: "string"
      price:
        type: "integer"
      total_price:
        type: "integer"
      quantity:
        type: "integer"
      store_id:
        type: "integer"
      user_id:
        type: "integer"
      checked_out:
        type: "boolean"
      served:
        type: "boolean"
...
  Category:
    type: "object"
    properties:
      id:
        type: "string"
      name:
        type: "string" 
      picture:
        type: "string"
      description:
        type: "string"
  Product:
    type: "object"
    properties:
      id:
        type: "string"
      name:
        type: "string" 
      picture:
        type: "string"
      description:
        type: "string"
      price:
        type: "integer"
      gem:
        type: "boolean"
securityDefinitions:
  ApiKeyAuth:
    type: "apiKey"
    in: header
    name: "x-access-token"